Moved commandline parser into cli dir, fixed examples
